Only insert boilerplate AS when in actionscript-mode.
The previous implementation relied on a `filename` variable being set;
this could be a source of bugs if called from an unexpected location.

Calling `insert-flash-boilerplate` from `find-file-not-found-hooks` was
too global and was causing this function to be called from unexpected
locations. Now it is called as part of the `actionscript-mode-hook` on
empty buffers.